+++ src/Screens/CityManager.cs
@@ -33,7 +33,6 @@ namespace CivOne.Screens
 		private readonly bool _viewCity;
 		
 		private bool _update = true;
-		private bool _redraw = false;
 		private bool _mouseDown = false;
 
 		private int ExtraWidth => (Width - 320);
@@ -50,10 +49,9 @@ namespace CivOne.Screens
 			Destroy();
 		}
 		
-		private void DrawLayer(IScreen layer, uint gameTick, int x, int y)
+		private void DrawLayer(IScreen layer, int x, int y)
 		{
 			if (layer == null) return;
-			if (!layer.Update(gameTick) && !_redraw) return;
 			this.AddLayer(layer, x, y);
 		}
 		
@@ -70,14 +68,14 @@ namespace CivOne.Screens
 
 			if (_update)
 			{
-				DrawLayer(_cityHeader, gameTick, 2, 1);
-				DrawLayer(_cityResources, gameTick, 2, 23);
-				DrawLayer(_cityUnits, gameTick, 2, 67);
-				DrawLayer(_cityMap, gameTick, 127 + ExtraLeft, 23);
-				DrawLayer(_cityBuildings, gameTick, 211 + ExtraLeft, 1);
-				DrawLayer(_cityFoodStorage, gameTick, 2, 106);
-				DrawLayer(_cityInfo, gameTick, 95 + ExtraLeft, 106);
-				DrawLayer(_cityProduction, gameTick, 230 + ExtraLeft, 99);
+				DrawLayer(_cityHeader, 2, 1);
+				DrawLayer(_cityResources, 2, 23);
+				DrawLayer(_cityUnits, 2, 67);
+				DrawLayer(_cityMap, 127 + ExtraLeft, 23);
+				DrawLayer(_cityBuildings, 211 + ExtraLeft, 1);
+				DrawLayer(_cityFoodStorage, 2, 106);
+				DrawLayer(_cityInfo, 95 + ExtraLeft, 106);
+				DrawLayer(_cityProduction, 230 + ExtraLeft, 99);
 
 				DrawButton("Rename", 9, 1, 231 + ExtraLeft, (Height - 10), 42);
 				DrawButton("Exit", 12, 4, (Width - 36), (Height - 10), 33);